409 Conflitct response

+++ mepserver/mp1/application_support/controllers/app_confirmation_controller.py
@@ -1,11 +1,12 @@
 import json
 import sys
-
+import jsonschema
 import cherrypy
 
 sys.path.append("../../")
 from mp1.models import *
 from mp1.enums import IndicationType
+
 
 
 class ApplicationConfirmationController:
@@ -49,6 +50,7 @@ class ApplicationConfirmationController:
             return None
 
     @cherrypy.tools.json_in()
+    @json_out(cls=NestedEncoder)
     def application_confirm_termination(self, appInstanceId: str):
         """
         This method is used to confirm the application level termination of an application instance.
@@ -56,17 +58,20 @@ class ApplicationConfirmationController:
         :type appInstanceId: str
 
         HTTP STATUS CODE: 204, 401, 403, 404, 409, 429
+
+        Section 7.2.11.3.4
         """
-        # TODO notification to MEC app instance indicating it will be terminated
-        # or stopped soon. In the notification 
 
         try:
             appTerminationConfirmation = AppTerminationConfirmation.from_json(
                 cherrypy.request.json
             )
-
-        except:
-            print("Exception launched!")
+        except jsonschema.exceptions.ValidationError:
+            error_msg = "Request body must have only the 'operationAction' "   \
+                        "attribute which can only have one of two possible "   \
+                        "values: TERMINATING or STOPPING." 
+            error = BadRequest(error_msg)
+            return error.message()
 
         query_appStatus = dict(appInstanceId=appInstanceId)
         appStatus = cherrypy.thread_data.db.query_col(
@@ -76,20 +81,49 @@ class ApplicationConfirmationController:
         )
 
         if appStatus:
+            # MEC Platform has already notified MEC App that it will be 
+            # terminated or stopped soon (AppTerminationNotification). 
+            # At that point, in appStatus collection, app's indication attribute
+            # is changed to "TERMINATING" or "STOPPING". This attribute must match
+            # that sent in AppTerminationConfirmation raising an error otherwise.
+            operationAction = str(appTerminationConfirmation.operationAction)
+            if appStatus['indication'] != operationAction:
+                error_msg = f"There is no {operationAction.lower()} ongoing."
+                error = Conflict(error_msg)
+                return error.message()
+
             if len(appStatus['services']) > 0:
+                # seek for app services
                 serv_lst = []
                 for serv in appStatus['services']:
                     serv_lst.append(serv['serInstanceId'])
                 
+
+                # TODO deactivate traffic rules
+
+                # TODO deactivate dns rules
+
+                # TODO remove the MEC app instance from the list of instances to
+                # be notified about service availability (subsriptions)
+
+                # TODO sending service availability notification to the MEC apps
+                # that consumes the services produced by the terminating/stopping
+                # MEC app instance
+
+
+                # app services removal from services collection
                 if len(serv_lst) > 0:
-                    in_serv_lst = {}
+                    in_serv_lst = dict()
                     in_serv_lst['$in'] = serv_lst
                     query_services = dict(serInstanceId=in_serv_lst)
-                    print("\nQUERYYYY")
-                    pprint.pprint(query_services)
-                    # cherrypy.thread_data.db.remove("services", query_services)
                     cherrypy.thread_data.db.remove_many('services', query_services)
+            # app removal from appStatus collection
             cherrypy.thread_data.db.remove("appStatus", query_appStatus)
+        else:
+            error_msg = "The application instance resource is not instantiated."
+            error = Conflict(error_msg)
+            return error.message()
+
         
         cherrypy.response.status = 204
